جاري العمل على المرحلة الثانية
تم نقل الكائن الى ملف منفصل

تم تعديل الرقم في ملف رمز الى long double
