+++ winBuild/Alif5.vcxproj.filters
@@ -64,6 +64,9 @@
     <ClInclude Include="..\source\Error.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\source\Object.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="Alif5.rc">
